+++ pytest/tests/sanity/upgradable.py
@@ -17,7 +17,7 @@ sys.path.append('lib')
 import branches
 import cluster
 from utils import wait_for_blocks_or_timeout, load_binary_file
-from transaction import sign_deploy_contract_tx, sign_function_call_tx
+from transaction import sign_deploy_contract_tx, sign_function_call_tx, sign_payment_tx
 
 
 def main():
@@ -76,10 +76,27 @@ def main():
 
     # write some random value
     tx = sign_function_call_tx(nodes[0].signer_key, nodes[0].signer_key.account_id,
-                               'write_random_value', [], 100000000000, 0, 2,
+                               'write_random_value', [], 10**13, 0, 2,
                                base58.b58decode(hash.encode('utf8')))
     res = nodes[0].send_tx_and_wait(tx, timeout=20)
     assert 'error' not in res, res
+    assert 'Failure' not in res['result']['status'], res
+
+    # hex_account_id = (b"I'm hex!" * 4).hex()
+    hex_account_id = '49276d206865782149276d206865782149276d206865782149276d2068657821'
+    tx = sign_payment_tx(key=nodes[0].signer_key,
+                         to=hex_account_id,
+                         amount=10 ** 25,
+                         nonce=3,
+                         blockHash=base58.b58decode(hash.encode('utf8')))
+    res = nodes[0].send_tx_and_wait(tx, timeout=20)
+    # Account doesn't exist
+    assert 'error' not in res, res
+    assert 'Failure' in res['result']['status'], res
+
+    # No account
+    res = nodes[0].get_account(hex_account_id)
+    assert 'error' in res, res
 
     wait_for_blocks_or_timeout(nodes[0], 20, 120)
 
@@ -94,20 +111,36 @@ def main():
     status3 = nodes[3].get_status()
     protocol_version = status0['protocol_version']
     latest_protocol_version = status3["latest_protocol_version"]
-    assert protocol_version == latest_protocol_version,\
-           "Latest protocol version %d should match active protocol version %d" % (latest_protocol_version, protocol_version)
+    assert protocol_version == latest_protocol_version, \
+        "Latest protocol version %d should match active protocol version %d" % (latest_protocol_version, protocol_version)
 
     gas_price = nodes[0].json_rpc('gas_price', [None])
     gas_price = int(gas_price['result']['gas_price'])
     assert gas_price < 1000000000, gas_price
     assert gas_price > 100000000, gas_price
 
+    hash = status0['sync_info']['latest_block_hash']
+
     # write some random value again
     tx = sign_function_call_tx(nodes[0].signer_key, nodes[0].signer_key.account_id,
-                               'write_random_value', [], 100000000000, 0, 3,
+                               'write_random_value', [], 10**13, 0, 4,
                                base58.b58decode(hash.encode('utf8')))
     res = nodes[0].send_tx_and_wait(tx, timeout=20)
     assert 'error' not in res, res
+    assert 'Failure' not in res['result']['status'], res
+
+    tx = sign_payment_tx(key=nodes[0].signer_key,
+                         to=hex_account_id,
+                         amount=10 ** 25,
+                         nonce=5,
+                         blockHash=base58.b58decode(hash.encode('utf8')))
+    res = nodes[0].send_tx_and_wait(tx, timeout=20)
+    # Successfully created a new account on transfer to hex
+    assert 'error' not in res, res
+    assert 'Failure' not in res['result']['status'], res
+
+    hex_account_balance = int(nodes[0].get_account(hex_account_id)['result']['amount'])
+    assert hex_account_balance == 10 ** 25
 
 if __name__ == "__main__":
     main()